deep copy model

+++ iisc_codes/q2/center_loss.py
@@ -31,7 +31,7 @@ class CenterLoss(nn.Module):
         batch_size = x.size(0)
         distmat = torch.pow(x, 2).sum(dim=1, keepdim=True).expand(batch_size, self.num_classes) + \
                   torch.pow(self.centers, 2).sum(dim=1, keepdim=True).expand(self.num_classes, batch_size).t()
-        distmat.addmm_(1, -2, x, self.centers.t())
+        distmat.addmm_(1, -2, x, self.centers.t().double())
 
         classes = torch.arange(self.num_classes).long()
         if self.use_gpu: classes = classes.cuda()
